% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checksums.R
\name{createMd5Sums}
\alias{createMd5Sums}
\alias{md5SumsChanged}
\title{Utilities for Checksums}
\usage{
createMd5Sums(modulePkg)

md5SumsChanged(modulePkg, moduleLibrary)
}
\arguments{
\item{modulePkg}{the path to a JASP module.}

\item{moduleLibrary}{the build folder where the JASP module is installed.}
}
\value{
`createMd5Sums` returns a named vector with md5sum for each file name. The names of the vector are the file names.
`md5SumsChanged` returns TRUE if the checksums are different from the local files and FALSE otherwise.
}
\description{
Utilities for Checksums
}
