% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findChangedModules.R
\name{writeModuleStatusObject}
\alias{writeModuleStatusObject}
\title{Write Status Object of JASP Modules}
\usage{
writeModuleStatusObject(jaspRoot)
}
\arguments{
\item{jaspRoot}{the root of the jasp-desktop github repository}
}
\description{
Write Status Object of JASP Modules
}
